# backend/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Global Multi-Entity ERP API",
    "version": "0.4.0",
    "description": "API contract for global multi-entity accounting endpoints under /api/v1.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.global-ledger.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/security/roles": {
      "get": {
        "tags": [
          "Security"
        ],
        "operationId": "listRoles",
        "summary": "List roles",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "includePermissions",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Security"
        ],
        "operationId": "upsertRole",
        "summary": "Create or update role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/CreatedResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/security/roles/{roleId}/permissions": {
      "get": {
        "tags": [
          "Security"
        ],
        "operationId": "listRolePermissions",
        "summary": "List permissions of role",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "description": "Role identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Security"
        ],
        "operationId": "assignRolePermissions",
        "summary": "Assign permissions to role",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "description": "Role identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionsInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permissions assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/security/role-assignments": {
      "get": {
        "tags": [
          "Security"
        ],
        "operationId": "listRoleAssignments",
        "summary": "List role assignments",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "userId",
            "required": false,
            "description": "User identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "roleId",
            "required": false,
            "description": "Role identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "scopeId",
            "required": false,
            "description": "Scope identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "scopeType",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "TENANT",
                "GROUP",
                "COUNTRY",
                "LEGAL_ENTITY",
                "OPERATING_UNIT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role assignment list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Security"
        ],
        "operationId": "assignRoleToUserScope",
        "summary": "Assign role to user scope",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/OkResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/security/role-assignments/{assignmentId}": {
      "delete": {
        "tags": [
          "Security"
        ],
        "operationId": "deleteRoleAssignment",
        "summary": "Delete role assignment",
        "parameters": [
          {
            "in": "path",
            "name": "assignmentId",
            "required": true,
            "description": "Assignment identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role assignment deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/security/permissions": {
      "get": {
        "tags": [
          "Security"
        ],
        "operationId": "listPermissions",
        "summary": "List permissions",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "CreatedResponse": {
        "description": "Created",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Created"
            }
          }
        }
      },
      "OkResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Ok"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "AnyObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "RoleInput": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "isSystem": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "RolePermissionsInput": {
        "type": "object",
        "properties": {
          "permissionCodes": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissionCodes"
        ]
      },
      "RolePermissionsResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "roleId": {
            "type": "integer",
            "minimum": 1
          },
          "assignedPermissionCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "ok",
          "roleId",
          "assignedPermissionCount"
        ]
      },
...
    }
  }
}