# backend/src/routes/security.js
import express from "express";
import { query } from "../db.js";
import { assertScopeAccess, requirePermission } from "../middleware/rbac.js";
import { logRbacAuditEvent } from "../audit/rbacAuditLogger.js";
import {
  asyncHandler,
  assertRequiredFields,
  badRequest,
  parsePositiveInt,
  resolveTenantId,
} from "./_utils.js";

const router = express.Router();

function parseBoolean(value) {
  if (typeof value === "boolean") {
    return value;
  }
  if (typeof value !== "string") {
    return false;
  }
  const normalized = value.trim().toLowerCase();
  return normalized === "true" || normalized === "1" || normalized === "yes";
}

const VALID_SCOPE_TYPES = new Set([
  "TENANT",
  "GROUP",
  "COUNTRY",
  "LEGAL_ENTITY",
  "OPERATING_UNIT",
]);

const VALID_EFFECTS = new Set(["ALLOW", "DENY"]);

function normalizeScopeType(value) {
  const scopeType = String(value || "").toUpperCase();
  if (!VALID_SCOPE_TYPES.has(scopeType)) {
    throw badRequest(
      "scopeType must be one of TENANT, GROUP, COUNTRY, LEGAL_ENTITY, OPERATING_UNIT"
    );
  }
  return scopeType;
}

function normalizeEffect(value, fallback = "ALLOW") {
  const effect = String(value || fallback).toUpperCase();
  if (!VALID_EFFECTS.has(effect)) {
    throw badRequest("effect must be ALLOW or DENY");
  }
  return effect;
}

async function getRoleForTenant(roleId, tenantId) {
  const roleResult = await query(
    `SELECT id, tenant_id, code, name, is_system
     FROM roles
     WHERE id = ?
       AND tenant_id = ?
     LIMIT 1`,
    [roleId, tenantId]
  );
  return roleResult.rows[0] || null;
}

async function upsertPermissionAndGetId(permissionCode) {
  await query(
    `INSERT INTO permissions (code, description)
     VALUES (?, ?)
     ON DUPLICATE KEY UPDATE
       description = VALUES(description)`,
    [permissionCode, permissionCode]
  );

  const permissionResult = await query(
    `SELECT id FROM permissions WHERE code = ? LIMIT 1`,
    [permissionCode]
  );
  const permissionId = permissionResult.rows[0]?.id;
  if (!permissionId) {
    throw new Error(`Permission lookup failed for ${permissionCode}`);
  }
  return permissionId;
}

router.get(
  "/permissions",
  requirePermission("security.permission.read"),
  asyncHandler(async (req, res) => {
    const q = req.query.q ? String(req.query.q).trim() : null;
    const params = [];
    let whereClause = "";

    if (q) {
      whereClause = "WHERE code LIKE ? OR description LIKE ?";
      params.push(`%${q}%`, `%${q}%`);
    }

    const result = await query(
      `SELECT id, code, description
       FROM permissions
       ${whereClause}
       ORDER BY code`,
      params
    );

    return res.json({
      rows: result.rows,
    });
  })
);

router.get(
  "/users",
  requirePermission("security.role_assignment.read"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const q = req.query.q ? String(req.query.q).trim() : null;
    const conditions = ["tenant_id = ?"];
    const params = [tenantId];

    if (q) {
      conditions.push("(email LIKE ? OR name LIKE ?)");
      params.push(`%${q}%`, `%${q}%`);
    }

    const result = await query(
      `SELECT id, email, name, status, created_at
       FROM users
       WHERE ${conditions.join(" AND ")}
       ORDER BY name, email
       LIMIT 200`,
      params
    );

    return res.json({
      tenantId,
      rows: result.rows,
    });
  })
);

router.get(
  "/roles",
  requirePermission("security.role.read"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const includePermissions = parseBoolean(req.query.includePermissions);
    const result = await query(
      `SELECT id, tenant_id, code, name, is_system, created_at
       FROM roles
       WHERE tenant_id = ?
       ORDER BY code`,
      [tenantId]
    );

    const rows = result.rows || [];
    if (!includePermissions || rows.length === 0) {
      return res.json({ tenantId, rows });
    }

    const roleIds = rows.map((row) => row.id);
    const permissionResult = await query(
      `SELECT rp.role_id, p.code
       FROM role_permissions rp
       JOIN permissions p ON p.id = rp.permission_id
       WHERE rp.role_id IN (${roleIds.map(() => "?").join(", ")})
       ORDER BY rp.role_id, p.code`,
      roleIds
    );

    const codesByRoleId = new Map();
    for (const row of permissionResult.rows) {
      if (!codesByRoleId.has(row.role_id)) {
        codesByRoleId.set(row.role_id, []);
      }
      codesByRoleId.get(row.role_id).push(row.code);
    }

    const enriched = rows.map((row) => ({
      ...row,
      permissionCodes: codesByRoleId.get(row.id) || [],
    }));

    return res.json({
      tenantId,
      rows: enriched,
    });
  })
);

router.post(
  "/roles",
  requirePermission("security.role.upsert"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    assertRequiredFields(req.body, ["code", "name"]);
    const code = String(req.body.code).trim();
    const name = String(req.body.name).trim();
    const isSystem = Boolean(req.body.isSystem);

    const existingRoleResult = await query(
      `SELECT id
       FROM roles
       WHERE tenant_id = ?
         AND code = ?
       LIMIT 1`,
      [tenantId, code]
    );
    const existingRoleId = parsePositiveInt(existingRoleResult.rows[0]?.id);

    const result = await query(
      `INSERT INTO roles (tenant_id, code, name, is_system)
       VALUES (?, ?, ?, ?)
       ON DUPLICATE KEY UPDATE
         name = VALUES(name),
         is_system = VALUES(is_system)`,
      [tenantId, code, name, isSystem]
    );

    const roleId = result.rows.insertId || existingRoleId || null;
    const wasCreated = !existingRoleId && Boolean(result.rows.insertId);

    if (wasCreated) {
      await logRbacAuditEvent(req, {
        tenantId,
        action: "role.create",
        resourceType: "role",
        resourceId: roleId,
        payload: {
          code,
          name,
          isSystem,
        },
      });
    }

    return res.status(201).json({ ok: true, id: roleId });
  })
);

router.get(
  "/roles/:roleId/permissions",
  requirePermission("security.role.read"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const roleId = parsePositiveInt(req.params.roleId);
    if (!roleId) {
      throw badRequest("roleId must be a positive integer");
    }

    const role = await getRoleForTenant(roleId, tenantId);
    if (!role) {
      throw badRequest("Role not found");
    }

    const permissionResult = await query(
      `SELECT p.id, p.code, p.description
       FROM role_permissions rp
       JOIN permissions p ON p.id = rp.permission_id
       WHERE rp.role_id = ?
       ORDER BY p.code`,
      [roleId]
    );

    return res.json({
      role,
      permissions: permissionResult.rows,
    });
  })
);

router.post(
  "/roles/:roleId/permissions",
  requirePermission("security.role_permissions.assign"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const roleId = parsePositiveInt(req.params.roleId);
    if (!roleId) {
      throw badRequest("roleId must be a positive integer");
    }

    const role = await getRoleForTenant(roleId, tenantId);
    if (!role) {
      throw badRequest("Role not found");
    }

    const permissionCodes = Array.isArray(req.body?.permissionCodes)
      ? req.body.permissionCodes.map((code) => String(code).trim()).filter(Boolean)
      : [];

    if (permissionCodes.length === 0) {
      throw badRequest("permissionCodes must be a non-empty array");
    }

    for (const permissionCode of permissionCodes) {
      const permissionId = await upsertPermissionAndGetId(permissionCode);

      await query(
        `INSERT IGNORE INTO role_permissions (role_id, permission_id)
         VALUES (?, ?)`,
        [roleId, permissionId]
      );
    }

    return res.status(201).json({
      ok: true,
      roleId,
      assignedPermissionCount: permissionCodes.length,
    });
  })
);

router.put(
  "/roles/:roleId/permissions",
  requirePermission("security.role_permissions.assign"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const roleId = parsePositiveInt(req.params.roleId);
    if (!roleId) {
      throw badRequest("roleId must be a positive integer");
    }

    const role = await getRoleForTenant(roleId, tenantId);
    if (!role) {
      throw badRequest("Role not found");
    }

    const permissionCodesRaw = Array.isArray(req.body?.permissionCodes)
      ? req.body.permissionCodes
      : null;
    if (!permissionCodesRaw) {
      throw badRequest("permissionCodes must be an array");
    }

    const normalizedPermissionCodes = Array.from(
      new Set(permissionCodesRaw.map((code) => String(code).trim()).filter(Boolean))
    );

    const beforeResult = await query(
      `SELECT p.code
       FROM role_permissions rp
       JOIN permissions p ON p.id = rp.permission_id
       WHERE rp.role_id = ?
       ORDER BY p.code`,
      [roleId]
    );
    const beforeCodes = beforeResult.rows.map((row) => row.code);

    await query(`DELETE FROM role_permissions WHERE role_id = ?`, [roleId]);

    for (const permissionCode of normalizedPermissionCodes) {
      const permissionId = await upsertPermissionAndGetId(permissionCode);
      await query(
        `INSERT IGNORE INTO role_permissions (role_id, permission_id)
         VALUES (?, ?)`,
        [roleId, permissionId]
      );
    }

    await logRbacAuditEvent(req, {
      tenantId,
      action: "role.permission.replace",
      resourceType: "role",
      resourceId: roleId,
      scopeType: "TENANT",
      scopeId: tenantId,
      payload: {
        roleCode: role.code,
        beforePermissionCodes: beforeCodes,
        afterPermissionCodes: normalizedPermissionCodes,
      },
    });

    return res.json({
      ok: true,
      roleId,
      permissionCount: normalizedPermissionCodes.length,
    });
  })
);

router.get(
  "/role-assignments",
  requirePermission("security.role_assignment.read"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const userId = parsePositiveInt(req.query.userId);
    const roleId = parsePositiveInt(req.query.roleId);
    const scopeId = parsePositiveInt(req.query.scopeId);
    const scopeType = req.query.scopeType
      ? String(req.query.scopeType).toUpperCase()
      : null;

    const conditions = ["urs.tenant_id = ?"];
    const params = [tenantId];

    if (userId) {
      conditions.push("urs.user_id = ?");
      params.push(userId);
    }
    if (roleId) {
      conditions.push("urs.role_id = ?");
      params.push(roleId);
    }
    if (scopeType) {
      conditions.push("urs.scope_type = ?");
      params.push(scopeType);
    }
    if (scopeId) {
      conditions.push("urs.scope_id = ?");
      params.push(scopeId);
    }

    const result = await query(
      `SELECT
         urs.id,
         urs.user_id,
         u.email AS user_email,
         u.name AS user_name,
         urs.role_id,
         r.code AS role_code,
         r.name AS role_name,
         urs.scope_type,
         urs.scope_id,
         urs.effect,
         urs.created_at
       FROM user_role_scopes urs
       JOIN users u ON u.id = urs.user_id
       JOIN roles r ON r.id = urs.role_id
       WHERE ${conditions.join(" AND ")}
       ORDER BY urs.id DESC`,
      params
    );

    return res.json({
      tenantId,
      rows: result.rows,
    });
  })
);

router.post(
  "/role-assignments",
  requirePermission("security.role_assignment.upsert", {
    resolveScope: (req, tenantId) => {
      const scopeType = String(req.body?.scopeType || "TENANT").toUpperCase();
      const scopeId =
        parsePositiveInt(req.body?.scopeId) || parsePositiveInt(tenantId);
      return { scopeType, scopeId };
    },
  }),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    assertRequiredFields(req.body, [
      "userId",
      "roleId",
      "scopeType",
      "scopeId",
      "effect",
    ]);

    const userId = parsePositiveInt(req.body.userId);
    const roleId = parsePositiveInt(req.body.roleId);
    const scopeId = parsePositiveInt(req.body.scopeId);
    const scopeType = normalizeScopeType(req.body.scopeType);
    const effect = normalizeEffect(req.body.effect);

    if (!userId || !roleId || !scopeId) {
      throw badRequest("userId, roleId and scopeId must be positive integers");
    }

    const existingResult = await query(
      `SELECT id, effect
       FROM user_role_scopes
       WHERE tenant_id = ?
         AND user_id = ?
         AND role_id = ?
         AND scope_type = ?
         AND scope_id = ?
       LIMIT 1`,
      [tenantId, userId, roleId, scopeType, scopeId]
    );
    const existingAssignmentId = parsePositiveInt(existingResult.rows[0]?.id);

    await query(
      `INSERT INTO user_role_scopes (
          tenant_id, user_id, role_id, scope_type, scope_id, effect
        )
       VALUES (?, ?, ?, ?, ?, ?)
       ON DUPLICATE KEY UPDATE
         effect = VALUES(effect)`,
      [tenantId, userId, roleId, scopeType, scopeId, effect]
    );

    const currentResult = await query(
      `SELECT id
       FROM user_role_scopes
       WHERE tenant_id = ?
         AND user_id = ?
         AND role_id = ?
         AND scope_type = ?
         AND scope_id = ?
       LIMIT 1`,
      [tenantId, userId, roleId, scopeType, scopeId]
    );
    const currentAssignmentId = parsePositiveInt(currentResult.rows[0]?.id);

    if (!existingAssignmentId) {
      await logRbacAuditEvent(req, {
        tenantId,
        targetUserId: userId,
        action: "assignment.create",
        resourceType: "user_role_scope",
        scopeType,
        scopeId,
        payload: {
          userId,
          roleId,
          scopeType,
          scopeId,
          effect,
        },
      });
    }

    return res.status(201).json({
      ok: true,
      created: !existingAssignmentId,
      assignmentId: currentAssignmentId || existingAssignmentId || null,
    });
  })
);

router.put(
  "/role-assignments/:assignmentId/scope",
  requirePermission("security.role_assignment.upsert", {
    resolveScope: (req, tenantId) => {
      const scopeType = String(req.body?.scopeType || "TENANT").toUpperCase();
      const scopeId =
        parsePositiveInt(req.body?.scopeId) || parsePositiveInt(tenantId);
      return { scopeType, scopeId };
    },
  }),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const assignmentId = parsePositiveInt(req.params.assignmentId);
    if (!assignmentId) {
      throw badRequest("assignmentId must be a positive integer");
    }

    assertRequiredFields(req.body, ["scopeType", "scopeId", "effect"]);

    const scopeType = normalizeScopeType(req.body.scopeType);
    const scopeId = parsePositiveInt(req.body.scopeId);
    const effect = normalizeEffect(req.body.effect);

    if (!scopeId) {
      throw badRequest("scopeId must be a positive integer");
    }

    const assignmentResult = await query(
      `SELECT id, user_id, role_id, scope_type, scope_id, effect
       FROM user_role_scopes
       WHERE id = ?
         AND tenant_id = ?
       LIMIT 1`,
      [assignmentId, tenantId]
    );
    const assignment = assignmentResult.rows[0];
    if (!assignment) {
      throw badRequest("Role assignment not found");
    }

    const oldScopeType = String(assignment.scope_type || "").toLowerCase();
    const oldScopeId = parsePositiveInt(assignment.scope_id);
    if (oldScopeType && oldScopeType !== "tenant" && oldScopeId) {
      assertScopeAccess(req, oldScopeType, oldScopeId, "existing scope");
    }

    await query(
      `UPDATE user_role_scopes
       SET scope_type = ?,
           scope_id = ?,
           effect = ?
       WHERE id = ?
         AND tenant_id = ?`,
      [scopeType, scopeId, effect, assignmentId, tenantId]
    );

    await logRbacAuditEvent(req, {
      tenantId,
      targetUserId: parsePositiveInt(assignment.user_id),
      action: "assignment.scope_replace",
      resourceType: "user_role_scope",
      resourceId: assignmentId,
      scopeType,
      scopeId,
      payload: {
        assignmentId,
        userId: parsePositiveInt(assignment.user_id),
        roleId: parsePositiveInt(assignment.role_id),
        before: {
          scopeType: String(assignment.scope_type || "").toUpperCase(),
          scopeId: parsePositiveInt(assignment.scope_id),
          effect: String(assignment.effect || "").toUpperCase(),
        },
        after: {
          scopeType,
          scopeId,
          effect,
        },
      },
    });

    return res.json({
      ok: true,
      assignmentId,
      scopeType,
      scopeId,
      effect,
    });
  })
);

router.delete(
  "/role-assignments/:assignmentId",
  requirePermission("security.role_assignment.upsert"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const assignmentId = parsePositiveInt(req.params.assignmentId);
    if (!assignmentId) {
      throw badRequest("assignmentId must be a positive integer");
    }

    await query(
      `DELETE FROM user_role_scopes
       WHERE id = ?
         AND tenant_id = ?`,
      [assignmentId, tenantId]
    );

    return res.json({ ok: true });
  })
);

router.get(
  "/data-scopes",
  requirePermission("security.data_scope.read"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const userId = parsePositiveInt(req.query.userId);
    const scopeType = req.query.scopeType
      ? String(req.query.scopeType).toUpperCase()
      : null;
    const scopeId = parsePositiveInt(req.query.scopeId);

    const conditions = ["ds.tenant_id = ?"];
    const params = [tenantId];

    if (userId) {
      conditions.push("ds.user_id = ?");
      params.push(userId);
    }
    if (scopeType) {
      conditions.push("ds.scope_type = ?");
      params.push(scopeType);
    }
    if (scopeId) {
      conditions.push("ds.scope_id = ?");
      params.push(scopeId);
    }

    const result = await query(
      `SELECT
         ds.id,
         ds.tenant_id,
         ds.user_id,
         u.email AS user_email,
         u.name AS user_name,
         ds.scope_type,
         ds.scope_id,
         ds.effect,
         ds.created_by_user_id,
         ds.created_at,
         ds.updated_at
       FROM data_scopes ds
       JOIN users u ON u.id = ds.user_id
       WHERE ${conditions.join(" AND ")}
       ORDER BY ds.id DESC`,
      params
    );

    return res.json({
      tenantId,
      rows: result.rows,
    });
  })
);

router.post(
  "/data-scopes",
  requirePermission("security.data_scope.upsert", {
    resolveScope: (req, tenantId) => {
      const scopeType = String(req.body?.scopeType || "TENANT").toUpperCase();
      const scopeId =
        parsePositiveInt(req.body?.scopeId) || parsePositiveInt(tenantId);
      return { scopeType, scopeId };
    },
  }),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    assertRequiredFields(req.body, ["userId", "scopeType", "scopeId"]);
    const userId = parsePositiveInt(req.body.userId);
    const scopeType = normalizeScopeType(req.body.scopeType);
    const scopeId = parsePositiveInt(req.body.scopeId);
    const effect = normalizeEffect(req.body.effect);
    const createdByUserId = parsePositiveInt(req.user?.userId);

    if (!userId || !scopeId) {
      throw badRequest("userId and scopeId must be positive integers");
    }

    await query(
      `INSERT INTO data_scopes (
          tenant_id, user_id, scope_type, scope_id, effect, created_by_user_id
       )
       VALUES (?, ?, ?, ?, ?, ?)
       ON DUPLICATE KEY UPDATE
         effect = VALUES(effect),
         created_by_user_id = VALUES(created_by_user_id)`,
      [tenantId, userId, scopeType, scopeId, effect, createdByUserId]
    );

    return res.status(201).json({ ok: true });
  })
);

router.put(
  "/data-scopes/users/:userId/replace",
  requirePermission("security.data_scope.upsert"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const userId = parsePositiveInt(req.params.userId);
    if (!userId) {
      throw badRequest("userId must be a positive integer");
    }

    const scopes = Array.isArray(req.body?.scopes) ? req.body.scopes : null;
    if (!scopes) {
      throw badRequest("scopes must be an array");
    }

    const normalizedScopes = [];
    for (const scope of scopes) {
      const scopeType = normalizeScopeType(scope?.scopeType);
      const scopeId = parsePositiveInt(scope?.scopeId);
      const effect = normalizeEffect(scope?.effect);
      if (!scopeId) {
        throw badRequest("Each scope item requires a positive scopeId");
      }
      normalizedScopes.push({ scopeType, scopeId, effect });
    }

    const createdByUserId = parsePositiveInt(req.user?.userId);

    const beforeResult = await query(
      `SELECT scope_type, scope_id, effect
       FROM data_scopes
       WHERE tenant_id = ?
         AND user_id = ?
       ORDER BY scope_type, scope_id`,
      [tenantId, userId]
    );

    await query(
      `DELETE FROM data_scopes
       WHERE tenant_id = ?
         AND user_id = ?`,
      [tenantId, userId]
    );

    for (const scope of normalizedScopes) {
      await query(
        `INSERT INTO data_scopes (
            tenant_id, user_id, scope_type, scope_id, effect, created_by_user_id
         )
         VALUES (?, ?, ?, ?, ?, ?)`,
        [
          tenantId,
          userId,
          scope.scopeType,
          scope.scopeId,
          scope.effect,
          createdByUserId,
        ]
      );
    }

    await logRbacAuditEvent(req, {
      tenantId,
      targetUserId: userId,
      action: "assignment.scope_replace",
      resourceType: "data_scope",
      resourceId: userId,
      scopeType: "TENANT",
      scopeId: tenantId,
      payload: {
        userId,
        beforeScopes: beforeResult.rows,
        afterScopes: normalizedScopes,
      },
    });

    return res.json({
      ok: true,
      userId,
      scopeCount: normalizedScopes.length,
    });
  })
);

router.delete(
  "/data-scopes/:dataScopeId",
  requirePermission("security.data_scope.upsert"),
  asyncHandler(async (req, res) => {
    const tenantId = resolveTenantId(req);
    if (!tenantId) {
      throw badRequest("tenantId is required");
    }

    const dataScopeId = parsePositiveInt(req.params.dataScopeId);
    if (!dataScopeId) {
      throw badRequest("dataScopeId must be a positive integer");
    }

    await query(
      `DELETE FROM data_scopes
       WHERE id = ?
         AND tenant_id = ?`,
      [dataScopeId, tenantId]
    );

    return res.json({ ok: true });
  })
);

export default router;